+++ docker/clickhouse/init.sql
@@ -2,7 +2,7 @@ CREATE DATABASE IF NOT EXISTS dew;
 USE dew;
 
 CREATE TABLE data (
-    device_id UInt16,
+    device_id UInt64,
     indoor_temperature Float32,
     indoor_humidity Float32,
     indoor_dewpoint Float32,
