+++ rust/arDEWino-rs/src/models.rs
@@ -1,5 +1,5 @@
 pub struct IndoorSensorData {
-    pub device_id: u16,
+    pub device_id: u64,
     pub temperature: f32,
     pub humidity: f32,
 }
